+++ _Typings/gameApi/types/builtin.d.ts
@@ -2,7 +2,7 @@
 // For details: https://github.com/Withered-Flower-0422/BST/blob/main/LICENSE
 
 declare module "game:type" {
-    type BuiltinCollections =
+    type BuiltinCollection =
         | "Score"
         | "Cure"
         | "Recover"
@@ -15,5 +15,5 @@ declare module "game:type" {
         | "Key_Magenta"
         | "Key_Yellow"
 
-    type BuiltinVariables = "C_Key_Red" | "C_Key_Blue" | "C_Key_Green" | "C_Key_Cyan" | "C_Key_Magenta" | "C_Key_Yellow"
+    type BuiltinVariable = "C_Key_Red" | "C_Key_Blue" | "C_Key_Green" | "C_Key_Cyan" | "C_Key_Magenta" | "C_Key_Yellow"
 }